Separada la logida de JavaScript en un hook personalizado: useArticleLogic.js, ademas se configuro el archivo peticionesArticle.js porque daba falla al realizar las solicitudes post y patch
