+++ src/components/ArticleForm.jsx
@@ -1,10 +1,7 @@
-import { useEffect, useState } from 'react';
 import { Table } from './shared/Table';
 import { DynamicForm } from './shared/DinamicForm/DynamicForm';
-import { PlusIcon } from '@heroicons/react/24/solid';
-import { FaMinus } from "react-icons/fa6";
 import ItemFormPopup from '../components/shared/FormCreation/FormCreation';
-import {callApiArticle,createItemArticle ,updateItemArticle, deleteItemArticle} from '../axios/peticionesArticle'
+import { useArticleLogic } from './hooks/useArticleLogic';
 
 const ArticleField = {
   keys: ['id', 'title', 'description', 'image', 'createdAt', 'updatedAt'],
@@ -23,127 +20,27 @@ const Createfields = [
 ];
 
 export default function ArticleForm() {
-  const [showPopup, setShowPopup] = useState(false);
-  const [data, setData] = useState([]);
-  const [totalItems, setTotalItems] = useState(0);
-  const [totalPages, setTotalPages] = useState(1);
-  const [currentPage, setCurrentPage] = useState(1);
-  const [currentItem, setCurrentItem] = useState(null);
-  const [searchParams, setSearchParams] = useState({});
-  const [resetForm, setResetForm] = useState(false);
-  const [formAction, setFormAction] = useState(true)
 
-  useEffect(() => {
-    const fetchData = async () => {
-      try {
-        const response = await callApiArticle(currentPage, 5, searchParams);
-        setData(response.data);
-        setTotalItems(response.pagination.totalItems);
-        setTotalPages(response.pagination.pageCount);
-      } catch (error) {
-        console.error('Error fetching data:', error);
-      }
-    };
+  const {
+    actions,
+    data,
+    extraButtons,
+    handleFormSubmit,
+    handlePageChange,
+    resetForm,
+    searchFormSubmit,
+    showPopup,
+    totalItems,
+    totalPages,
+    currentItem,
+    closePopup,
+    currentPage,
+    formAction,
+    setCurrentItem
+  } = useArticleLogic()
 
-    fetchData();
-  }, [currentPage, searchParams]);
-
-  const openPopup = () => {
-    setShowPopup(true);
-  };
-
-  const closePopup = () => {
-    setShowPopup(false);
-    setCurrentItem(null);
-  };
-
-  const handleFormSubmit = async (event) => {
-
-    try {
-      event.preventDefault();
-      const formData = new FormData(event.target);
-      const formValues = Object.fromEntries(formData.entries());
-      (formAction ? await createItemArticle({ ...formValues }) : await updateItemArticle(currentItem.Id, { ...formValues }))
-      closePopup()
-      setCurrentPage(1);
-      const response = await callApiArticle(currentPage, 5, searchParams);
-      setData(response.data);
-    } catch (error) {
-      console.log('Ocurrio un error en el servidor' , error);
-    }
-
-  };
-
-  const searchFormSubmit = async (form) => {
-
-    console.log(form)
- 
-    setSearchParams(form);
-
-    const response = await callApiArticle(currentPage, 5, form);
-    setData(response.data);
-
-    //resetAllForms();
-  };
-
-  const handleCreate = () => {
-    setFormAction(true)
-    openPopup();
-  };
-
-  const handleEdit = (item) => {
-    console.log('Edit item:', item);
-    setCurrentItem(item);
-    setFormAction(false)
-    openPopup();
-  };
-
-  const handleDelete = async (item) => {
-    console.log('Delete item:', item);
-    try {
-      await deleteItemArticle(item.Id);
-      const response = await callApiArticle(currentPage, 5, searchParams);
-      setData(response.data);
-      setTotalItems(response.pagination.totalItems);
-      setTotalPages(response.pagination.pageCount);
-    } catch (error) {
-      console.error('Error deleting item:', error);
-    }
-  };
-
-  const handlePageChange = (newPage) => {
-    setCurrentPage(newPage);
-  };
-
-  const resetAllForms = () => {
-    setSearchParams({});
-    setCurrentPage(1);
-    setCurrentItem(null);
-    setResetForm(true); // Trigger form reset
-    setTimeout(() => setResetForm(false), 0); // Reset the flag
-  };
-
-  const extraButtons = [
-    {
-      label: 'Crear Article',
-      onClick: handleCreate,
-      className: 'bg-indigo-500 hover:bg-indigo-700 crear',
-      icon: PlusIcon,
-    },
-    {
-      label: 'Restablecer',
-      onClick: resetAllForms,
-      className: 'bg-red-500 hover:bg-red-700',
-      icon: FaMinus,
-    },
-  ];
-
-  const actions = [
-    { label: 'Editar', onClick: handleEdit },
-    { label: 'Eliminar', onClick: handleDelete },
-  ];
-  return(
-<div className="min-h-screen bg-white flex flex-col">
+  return (
+    <div className="min-h-screen bg-white flex flex-col">
       <DynamicForm fields={fields} onSubmit={searchFormSubmit} extraButtons={extraButtons} resetForm={resetForm} />
 
       {showPopup && (
